+++ src/sqlfluff/dialects/dialect_bigquery.py
@@ -18,7 +18,6 @@ from sqlfluff.core.parser import (
     Delimited,
     AnyNumberOf,
     KeywordSegment,
-    Indent,
     SymbolSegment,
     RegexLexer,
     StringLexer,
@@ -83,6 +82,11 @@ bigquery_dialect.add(
     RightArrowSegment=StringParser(
         "=>", SymbolSegment, name="right_arrow", type="right_arrow"
     ),
+    SelectClauseElementListGrammar=Delimited(
+        Ref("SelectClauseElementSegment"),
+        delimiter=Ref("CommaSegment"),
+        allow_trailing=True,
+    ),
 )
 
 
@@ -143,6 +147,18 @@ bigquery_dialect.sets("angle_bracket_pairs").update(
 )
 
 
+@bigquery_dialect.segment(replace=True)
+class SelectClauseModifierSegment(BaseSegment):
+    """Things that come after SELECT but before the columns."""
+
+    type = "select_clause_modifier"
+    match_grammar = Sequence(
+        # https://cloud.google.com/bigquery/docs/reference/standard-sql/query-syntax
+        Sequence("AS", OneOf("STRUCT", "VALUE"), optional=True),
+        OneOf("DISTINCT", "ALL", optional=True),
+    )
+
+
 # BigQuery allows functions in INTERVAL
 @bigquery_dialect.segment(replace=True)
 class IntervalExpressionSegment(BaseSegment):
@@ -153,30 +169,6 @@ class IntervalExpressionSegment(BaseSegment):
         "INTERVAL",
         Ref("ExpressionSegment"),
         OneOf(Ref("QuotedLiteralSegment"), Ref("DatetimeUnitSegment")),
-    )
-
-
-@bigquery_dialect.segment(replace=True)
-class SelectClauseSegment(ansi_dialect.get_segment("SelectClauseSegment")):  # type: ignore
-    """In BigQuery, select * as struct is valid."""
-
-    parse_grammar = Sequence(
-        "SELECT",
-        Ref("SelectClauseModifierSegment", optional=True),
-        Indent,
-        OneOf(
-            Sequence(
-                "AS",
-                "STRUCT",
-                Ref("StarSegment"),
-                Ref("StarModifierSegment", optional=True),
-            ),
-            Delimited(
-                Ref("SelectClauseElementSegment"),
-                delimiter=Ref("CommaSegment"),
-                allow_trailing=True,
-            ),
-        ),
     )